+++ bin/sync.sh
@@ -3,3 +3,4 @@
 git checkout v5-dev
 git pull osc v5-dev
 git pull origin v5-dev
+git pull gitcode v5-dev